Move and rename the plugins.
So that they take the exact same place of the previous ones.
